refactor: Fix thresholds slice filename typo
