+++ src/store/generator/slices/testOptions/index.ts
@@ -2,8 +2,8 @@ import { ImmerStateCreator } from '@/utils/typescript'
 
 import { LoadProfileStore, createLoadProfileSlice } from './loadProfile'
 import { LoadZoneStore, createLoadZoneSlice } from './loadZones'
-import { createThresholdSlice, ThresholdStore } from './thesholds'
 import { ThinkTimeStore, createThinkTimeSlice } from './thinkTime'
+import { createThresholdSlice, ThresholdStore } from './thresholds'
 
 export type TestOptionsStore = LoadProfileStore &
   ThinkTimeStore &
